[POC] Moving mutated php code to it's own file
