+++ tests/Fixtures/Autoloaded/PublicVisibility/pv-final-m0.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace PublicVisibilityFinal;
+
+class Test
+{
+    protected final function foo(int $param, $test = 1) : bool
+    {
+        echo 1;
+        return false;
+    }
+}
