+++ tests/Fixtures/Autoloaded/PublicVisibility/pv-non-abstract-in-abstract-class-m0.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace PublicVisibilityNonAbstractInAbstractClass;
+
+abstract class Test
+{
+    protected function foo(int $param, $test = 1) : bool
+    {
+        echo 1;
+        return false;
+    }
+}
