+++ tests/Fixtures/Autoloaded/PublicVisibility/pv-not-set-m0.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace PublicVisibilityNotSet;
+
+class Test
+{
+    protected function foo()
+    {
+    }
+}
