+++ tests/Fixtures/Autoloaded/PublicVisibility/pv-one-class-m0.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace PublicVisibilityOneClass;
+
+class Test
+{
+    protected function &foo(int $param, $test = 1) : bool
+    {
+        echo 1;
+        return false;
+    }
+}
