+++ tests/Fixtures/Autoloaded/PublicVisibility/pv-same-method-parent-m0.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace SameParent;
+
+class SameParent
+{
+    protected function foo()
+    {
+    }
+}
+class Child extends SameParent
+{
+    public function foo()
+    {
+    }
+}
