+++ tests/Fixtures/Autoloaded/PublicVisibility/pv-static-m0.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace PublicVisibilityStatic;
+
+class Test
+{
+    protected static function foo(int $param, $test = 1) : bool
+    {
+        echo 1;
+        return false;
+    }
+}
